requirements.txt
flask
boto3
Pillow
easyocr
